% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesAccessors.R
\docType{methods}
\name{accessOptions,CopyNumberBreakPoints-method}
\alias{accessOptions}
\alias{accessOptions,CopyNumberBreakPoints-method}
\title{Access Object Data}
\usage{
\S4method{accessOptions}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\description{
Access Object Data
}
\examples{
accessOptions( bp )
accessOptions( bp_genes )
accessOptions( bp_stats )
# --- Object data access ---
# This is an object of class "CopyNumberBreakPointGenes"
# callData( obj ) => returns feature call values
# segmentData( obj ) => returns feature segment values
# breakpointData( obj ) => returns feature breakpoint values
# sampleNames( obj ) => returns vector with sample names
}

