% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpFilter,CopyNumberBreakPoints-method}
\alias{bpFilter}
\alias{bpFilter,CopyNumberBreakPoints-method}
\title{Filter Breakpoints}
\usage{
\S4method{bpFilter}{CopyNumberBreakPoints}(object, filter = "CNA-ass",
  threshold = NULL)
}
\arguments{
\item{filter}{Type of filter}
}
\value{
Object of class \code{CopyNumberBreakPoints})
}
\description{
Removes breakpoints by a minimum <ADD_TERM!> threshold
}
\examples{
bp <- bpFilter( bp, filter = "deltaSeg", threshold = 0.2 )
}

