% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpGenes,CopyNumberBreakPointGenes-method}
\alias{bpGenes}
\alias{bpGenes,CopyNumberBreakPointGenes-method}
\title{bpGenes}
\usage{
\S4method{bpGenes}{CopyNumberBreakPointGenes}(object)
}
\arguments{
\item{object}{A "CopyNumberBreakPointGenes" object}

\item{geneAnnotations}{data.frame with gene annotations}
}
\value{
Object of same class of \code{object}.
}
\description{
Add description...
}
\examples{
bpGenes( breakPointsGenes, gene.annotations.hg19 )
}

