% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpPlot,CopyNumberBreakPoints-method}
\alias{bpPlot}
\alias{bpPlot,CopyNumberBreakPoints-method}
\title{bpPlot}
\usage{
\S4method{bpPlot}{CopyNumberBreakPoints}(object, plot.chr = NULL,
  plot.ylim = 15, fdr.threshold = 0.1, add.jitter = FALSE)
}
\arguments{
\item{object}{A "CopyNumberBreakPoint" object}

\item{plot.chr}{The chromosome(s) to plot}

\item{plot.ylim}{The max y}

\item{fdr.threshold}{The FDR threshold}
}
\value{
Nothing
}
\description{
Plots breakpoint frequencies per chromosome
}
\details{
The plot includes breakpoints locations and frequencies. Genes that are recurrently affected are labeled with their gene name.
}
\examples{
bpPlot( bp, c(1:4) )
}

