% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpStats,CopyNumberBreakPoints-method}
\alias{bpStats}
\alias{bpStats,CopyNumberBreakPoints-method}
\title{bpStats}
\usage{
\S4method{bpStats}{CopyNumberBreakPoints}(object, level = "gene",
  method = "BH", fdr.threshold = 1)
}
\arguments{
\item{object}{A "CopyNumberBreakPointGenes" object}

\item{level}{The level at which to operate, this can be either "gene" (adjusting for gene size) or "feature" (per probe/bin)}

\item{method}{The FDR correction method to apply. This can be "BH" (applies Benjamini Hochberg) or "Gilbert" (for dedicated Benjamini Hochberg)}
}
\value{
Object of same class of \code{object}.
}
\description{
Applies cohort based statistics to identify genes and/or features that are recurrently affected by breakpoint locations.
}
\details{
To be added...
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( dcopynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )
bp <- bpGenes( bp )
bp <- bpStats( bp )
}

