% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{breakpointsPerGene,CopyNumberBreakPointGenes-method}
\alias{breakpointsPerGene}
\alias{breakpointsPerGene,CopyNumberBreakPointGenes-method}
\title{Access Object breakpointsPerGene. This method returns a dataframe with breakpoints per gene.}
\usage{
\S4method{breakpointsPerGene}{CopyNumberBreakPointGenes}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\value{
a dataframe with breakpoints per gene
}
\description{
Access Object breakpointsPerGene. This method returns a dataframe with breakpoints per gene.
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )
bp <- bpGenes( bp )

breakpointsPerGene( bp )
}

