% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{callData,CopyNumberBreakPoints-method}
\alias{callData}
\alias{callData,CopyNumberBreakPoints-method}
\title{Access Object callData}
\usage{
\S4method{callData}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\description{
Access Object callData
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )
callData( bp )
}

