% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{copynumber.data.chr20}
\alias{copynumber.data.chr20}
\title{CGHcall Example copynumber data hg18 chr20}
\format{An object of class cghCall}
\usage{
data( copynumber.data.chr20 )
}
\value{
An object of class cghCall
}
\description{
A test dataset containing chromosome 20 copynumber data for the GeneBreak package (hg18 based).
}
\keyword{datasets}

