% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CopyNumberBreakPoints-class}
\alias{CopyNumberBreakPoints}
\alias{CopyNumberBreakPoints-class}
\title{An S4 class to represent a CopyNumberBreakPoints object.}
\description{
An S4 class to represent a CopyNumberBreakPoints object.
}
\section{Slots}{

\describe{
\item{\code{segmDiff}}{A matrix with breakpoints based on segment values}

\item{\code{callDiff}}{A matrix with breakpoints based on call values}

\item{\code{segments}}{A matrix with segmented copy number values}

\item{\code{calls}}{A matrix with copy number calls}

\item{\code{featureAnnotation}}{A dataframe with predefined information about the features (usually probes or bins)}

\item{\code{featureData}}{A dataframe with calculated information about the features (usually probes or bins)}
}}

