% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\name{createGeneAnnotations}
\alias{createGeneAnnotations}
\title{Build Gene Annotations}
\usage{
createGeneAnnotations(file, geneCount = NULL)
}
\arguments{
\item{file}{Path to a TAB delim file with gene annotations. Requires a header to be present with at least these fields: Gene, EnsID, Chromosome, Start, End}

\item{geneCount}{Only use first geneCount genes (for testing purposes)}
}
\value{
Object of class \code{CopyNumberBreakPoints})
}
\description{
Builds the geneAnnotation data.frame from a text file
}
\examples{
geneAnn <- createGeneAnnotations( "/path/to/annotation/file.tsv" )
}

