% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doSomething}
\alias{doSomething}
\title{Does something, nice huh!}
\usage{
doSomething(someObject, someFiles = NULL, someString = "bam")
}
\arguments{
\item{someObj}{A "BreakPointGenes" object.}

\item{y}{Empty.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Does something, nice huh!
}
\examples{
doSomething( bpGenes )
doSomething( bpGenes, nounou )
}

