% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ens.gene.ann.hg18.chr20}
\alias{ens.gene.ann.hg18.chr20}
\title{Gene Annotation hg18 chr20}
\format{A data.frame}
\usage{
data( ens.gene.ann.hg18.chr20 )
}
\description{
A test dataset containing the gene locations of chr20
 required for the BreakPointGenes package.
}
\details{
See help( ens.gene.ann.hg18 ) for more info on contents
}
\keyword{datasets}

