% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ens.gene.ann.hg18}
\alias{ens.gene.ann.hg18}
\title{Gene Annotation hg18}
\format{A data.frame}
\usage{
data( ens.gene.ann.hg18 )
head( ens.gene.ann.hg18 )
}
\description{
A dataset containing the gene locations
 required for the BreakPointGenes package.
}
\details{
Dataframe with 5 columns:
\itemize{
  \item Gene: ensembl gene name
  \item EnsID: ensembl gene id
  \item Chromosome: Genomic Chromosome
  \item Start: Genomic start of gene
  \item End: Genomic end of gene
}
}
\keyword{datasets}

