% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ens.gene.ann.hg19}
\alias{ens.gene.ann.hg19}
\title{Gene Annotation hg19}
\format{A data.frame}
\usage{
data( ens.gene.ann.hg19 )
}
\value{
data.frame
}
\description{
A dataset containing the gene locations based on human genome reference hg19 that was obtained from BioMart.
}
\details{
Dataframe with 5 columns:
\itemize{
  \item Gene: ensembl gene name
  \item EnsID: ensembl gene id
  \item Chromosome: Genomic Chromosome
  \item Start: Genomic start of gene
  \item End: Genomic end of gene
}
}
\keyword{datasets}

