% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesAccessors.R
\docType{methods}
\name{featureInfo,CopyNumberBreakPoints-method}
\alias{featureInfo}
\alias{featureInfo,CopyNumberBreakPoints-method}
\title{Gather Feature information}
\usage{
\S4method{featureInfo}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{of class \code{CopyNumberBreakPoints}}
}
\value{
data.frame
}
\description{
Gather Feature information
}
\examples{
featureInfo( bp )
}

