% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{featureNames,CopyNumberBreakPoints-method}
\alias{featureNames}
\alias{featureNames,CopyNumberBreakPoints-method}
\title{Access Object featureNames. This method returns a vector with feature names.}
\usage{
\S4method{featureNames}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\description{
Access Object featureNames. This method returns a vector with feature names.
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )

featureNames( bp )
}

