% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{gene.annotation.hg19.chr20}
\alias{gene.annotation.hg19.chr20}
\title{Gene Annotation hg19 chr20}
\format{A data.frame}
\usage{
data( gene.annotation.hg19.chr20 )
}
\description{
A dataset containing the gene locations
 required for the BreakPointGenes package.
}
\details{
See help( gene.annotation.hg19 ) for more info on contents
}
\keyword{datasets}

