% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{gene.annotation.hg19}
\alias{gene.annotation.hg19}
\title{Gene Annotation hg19}
\format{A data.frame}
\usage{
data( gene.annotation.hg19 )
}
\description{
A dataset containing the gene locations
 required for the BreakPointGenes package.
}
\details{
\itemize{
  \item Gene (gene name)
  \item EnsID (ensembl gene id)
  \item Chromosome
  \item Start (Genomic start of gene)
  \item End (Genomic end of gene)
}
}
\keyword{datasets}

