% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{geneChromosomes,CopyNumberBreakPointGenes-method}
\alias{geneChromosomes}
\alias{geneChromosomes,CopyNumberBreakPointGenes-method}
\title{Access Object geneChromosomes. This method returns a vector with gene chromosomes.}
\usage{
\S4method{geneChromosomes}{CopyNumberBreakPointGenes}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\value{
vector with gene chromosomes
}
\description{
Access Object geneChromosomes. This method returns a vector with gene chromosomes.
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )
bp <- bpGenes( bp )

geneChromosomes( bp )
}

