% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesAccessors.R
\docType{methods}
\name{geneInfo,CopyNumberBreakPointGenes-method}
\alias{geneInfo}
\alias{geneInfo,CopyNumberBreakPointGenes-method}
\title{Gather Gene information}
\usage{
\S4method{geneInfo}{CopyNumberBreakPointGenes}(object)
}
\arguments{
\item{object}{of class \code{CopyNumberBreakPointGenes}}
}
\value{
data.frame
}
\description{
Gather Gene information
}
\examples{
geneInfo( bp_genes )
}

