% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBP}
\alias{getBP}
\title{CopyNumber to BreakPoints}
\usage{
getBP(data, first.rm = TRUE)
}
\arguments{
\item{data}{A "CGHcall" object}

\item{first.rm}{Remove the first}
}
\value{
Object of class \code{CopyNumberBreakPoints}.
}
\description{
Builds the CopyNumberBreakPoints object with CGHcall data
}
\examples{
getBP( cghCallObj )
}

