% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\name{getBreakpoints}
\alias{getBreakpoints}
\title{CopyNumber to BreakPoints}
\usage{
getBreakpoints(data, first.rm = TRUE)
}
\arguments{
\item{data}{A "CGHcall" object}

\item{first.rm}{Remove the first breakpoint for each chromosome}
}
\value{
Object of class \code{CopyNumberBreakPoints}
}
\description{
Builds the CopyNumberBreakPoints object from CGHcall data
}
\examples{
data( copynumber.data.chr20 )
data( gene.annotation.hg19 )
breakpoints <- getBreakpoints( data = copynumber.data.chr20 )
}

