% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{LGG150.data}
\alias{LGG150.data}
\title{LGG150 cghCall object}
\format{An object of class "cghCall"}
\usage{
data( LGG150.data )
}
\description{
A dataset containing the copynumber, segment and call
 data for chromosomes 7, 8, 9 and 10 of a Low Grade
 Glioma sample (LGG150).
}
\details{
For more info on the cghCall class, see the package
 "CGHcall"
}
\keyword{datasets}

