% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{multiplyCounts,BreakPointGenes-method}
\alias{multiplyCounts}
\alias{multiplyCounts,BreakPointGenes-method}
\title{Multiplies the data}
\usage{
\S4method{multiplyCounts}{BreakPointGenes}(object, mFactor = 2)
}
\arguments{
\item{mFactor}{The multiply factor}

\item{data}{A "BreakPointGenes" object.}
}
\description{
Multiplies the data
}
\examples{
multiplyCounts( bpGenes, 10 )
}

