% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{namesFeatures,CopyNumberBreakPoints-method}
\alias{namesFeatures}
\alias{namesFeatures,CopyNumberBreakPoints-method}
\title{Access Object namesFeatures. This method returns a vector with feature names.}
\usage{
\S4method{namesFeatures}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\value{
a vector with feature names
}
\description{
Access Object namesFeatures. This method returns a vector with feature names.
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )

namesFeatures( bp )
}

