% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{PackageName}
\alias{PackageName}
\alias{PackageName-package}
\title{This is what it does.}
\description{
<packagename> allows you to do this and that
}
\details{
The only function you're likely to need from \pkg{BreakPointGenes} is
\code{\link{getBP}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

