% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDaShizzle}
\alias{plotDaShizzle}
\title{Plots da SHIZZLE....}
\usage{
plotDaShizzle(data, main = "Title", col = "red")
}
\arguments{
\item{x}{A "BreakPointGenes" object.}

\item{y}{Empty.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Plots da SHIZZLE....
}
\examples{
plotDaShizzle( bpGenes )
}

