% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readCGHcall}
\alias{readCGHcall}
\title{Creates a BreakPointGenes object from a CGHcall object}
\usage{
readCGHcall(cghCallObj, samples = c("A", "B"), fillInt = 9)
}
\arguments{
\item{someObj}{A "BreakPointGenes" object.}

\item{y}{Empty.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Creates a BreakPointGenes object from a CGHcall object
}
\examples{
plotDaShizzle( bpGenes )
}

