% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{recurrentGenes,CopyNumberBreakPointGenes-method}
\alias{recurrentGenes}
\alias{recurrentGenes,CopyNumberBreakPointGenes-method}
\title{Show recurrent genes}
\usage{
\S4method{recurrentGenes}{CopyNumberBreakPointGenes}(object,
  fdr.threshold = 0.1, summarize = TRUE, order.column = "FDR")
}
\arguments{
\item{object}{Output of bpStats(): a \code{CopyNumberBreakPointGenes} object}

\item{fdr.threshold}{Genes with lower FDR are returned}

\item{summarize}{If TRUE only certain columns are returned}

\item{order.column}{Name of the column to sort output on}
}
\value{
data.frame with recurrent genes
}
\description{
Show recurrent genes
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )
bp <- bpGenes( bp )
bp <- bpStats( bp )
recurrentGenes( bp )
}

