% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\name{runWorkflow}
\alias{runWorkflow}
\title{CopyNumber to BreakPoints}
\usage{
runWorkflow(object, geneAnnotation)
}
\arguments{
\item{data}{A "CGHcall" object}
}
\value{
Output of bpStats() an object of class \code{CopyNumberBreakPointGenes}.
}
\description{
Runs default settings for all workflow steps
}
\examples{
runWorkflow( cghCallObj )
}

