% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{sampleNames,CopyNumberBreakPoints-method}
\alias{sampleNames}
\alias{sampleNames,CopyNumberBreakPoints-method}
\title{Access Object sampleNames}
\usage{
\S4method{sampleNames}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\description{
Access Object sampleNames
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )
sampleNames( bp )
}

