% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sayHello}
\alias{sayHello}
\title{Sing a long!}
\usage{
sayHello(msg = "Is it me you're looking fooooor?",
  msg2 = "I can see it in your eyes..")
}
\arguments{
\item{msg}{Wha-ever}

\item{msg2}{Sure why not}
}
\value{
Returns NOTHING
}
\description{
Sing a long!
}
\examples{
sayHello()
sayHello( msg="First msg", msg2="First msg" )
}

