% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{segmentData,CopyNumberBreakPoints-method}
\alias{segmentData}
\alias{segmentData,CopyNumberBreakPoints-method}
\title{Access Object segmentData}
\usage{
\S4method{segmentData}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\description{
Access Object segmentData
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )
segmentData( bp )
}

