% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{addGeneAnnotation,CopyNumberBreakPoints-method}
\alias{addGeneAnnotation}
\alias{addGeneAnnotation,CopyNumberBreakPoints-method}
\title{addGeneAnnotation}
\usage{
\S4method{addGeneAnnotation}{CopyNumberBreakPoints}(object, geneAnnotation)
}
\arguments{
\item{object}{A "CopyNumberBreakPoints" object [output of getBreakpoints()]}
}
\value{
Object of same class of \code{object}.
}
\description{
Add description...
}
\examples{
addGeneAnnotation( breakPointsFiltered, geneAnnotation )
}

