% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addGeneAnnotation,CopyNumberBreakPoints-method}
\alias{addGeneAnnotation,CopyNumberBreakPoints-method}
\alias{geneToProbeAnn}
\title{geneToProbeAnn}
\usage{
\S4method{addGeneAnnotation}{CopyNumberBreakPoints}(object, geneAnnotations)
}
\arguments{
\item{object}{A "CopyNumberBreakPoints" object [output of getBP()]}
}
\value{
Object of same class of \code{object}.
}
\description{
\code{geneToProbeAnn} returns a list with.....
}
\examples{
getBP()
BPsample( bp, filter="none" )
}

