% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpFilter,CopyNumberBreakPoints-method}
\alias{bpFilter}
\alias{bpFilter,CopyNumberBreakPoints-method}
\title{Filter Breakpoints}
\usage{
\S4method{bpFilter}{CopyNumberBreakPoints}(object, filter = "CNA-ass",
  threshold = NULL)
}
\arguments{
\item{filter}{Type of filter. This can be either "deltaSeg", "deltaCall" or "CNA-ass".
\itemize{
  \item CNA-ass: copy number based
  \item deltaSeg:
  \item deltaCall:
} \cr}
}
\value{
Object of class \code{CopyNumberBreakPoints})
}
\description{
Selects breakpoints by criteria options.
}
\details{
Iets
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp, filter = "deltaSeg", threshold = 0.2 )
}

