% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpGenes,CopyNumberBreakPointGenes-method}
\alias{bpGenes}
\alias{bpGenes,CopyNumberBreakPointGenes-method}
\title{bpGenes}
\usage{
\S4method{bpGenes}{CopyNumberBreakPointGenes}(object)
}
\arguments{
\item{object}{A "CopyNumberBreakPointGenes" object}

\item{geneAnnotations}{data.frame with gene annotations}
}
\value{
Object of same class of \code{object}.
}
\description{
Indentifies genes involved in breakpoint locations.
}
\details{
This step requires gene annotation to be added to the breakpoints object (see ?addGeneAnnotation).
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )
bp <- bpGenes( bp )
}

