% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpPlot,CopyNumberBreakPoints-method}
\alias{bpPlot}
\alias{bpPlot,CopyNumberBreakPoints-method}
\title{bpPlot}
\usage{
\S4method{bpPlot}{CopyNumberBreakPoints}(object, plot.chr = NULL,
  plot.ylim = 15, fdr.threshold = 0.1)
}
\arguments{
\item{object}{A "CopyNumberBreakPoint" object}

\item{plot.chr}{The chromosome(s) to plot}

\item{plot.ylim}{The max y}

\item{fdr.threshold}{The FDR threshold}
}
\value{
Nothing
}
\description{
Add description...
}
\examples{
bpPlot( bp_stats, c(1,22) )
}

