% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\docType{methods}
\name{bpStats,CopyNumberBreakPoints-method}
\alias{bpStats}
\alias{bpStats,CopyNumberBreakPoints-method}
\title{bpStats}
\usage{
\S4method{bpStats}{CopyNumberBreakPoints}(object, level = "gene",
  method = "BH", fdr.threshold = 1)
}
\arguments{
\item{object}{A "CopyNumberBreakPointGenes" object}

\item{level}{The level ["gene" or "feature"]}

\item{method}{The method ["BH" or "Gilbert"]}
}
\value{
Object of same class of \code{object}.
}
\description{
Add description...
}
\examples{
bpStats( breakPointsGenes )
}

