% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{breakpointData,CopyNumberBreakPoints-method}
\alias{breakpointData}
\alias{breakpointData,CopyNumberBreakPoints-method}
\title{Access Object breakpointData. This method returns a dataframe with breakpoint values per feature.}
\usage{
\S4method{breakpointData}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\value{
a dataframe with breakpoint values
}
\description{
Access Object breakpointData. This method returns a dataframe with breakpoint values per feature.
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )

breakpointData( bp )
}

