% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{callData,CopyNumberBreakPoints-method}
\alias{callData}
\alias{callData,CopyNumberBreakPoints-method}
\title{Access Object callData. This method returns a dataframe with feature call values.}
\usage{
\S4method{callData}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\value{
a dataframe with feature call values
}
\description{
Access Object callData. This method returns a dataframe with feature call values.
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )

callData( bp )
}

