% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{copynumber.data.chr18}
\alias{copynumber.data.chr18}
\title{CGHcall CAIRO data hg19 chr18}
\format{A data.frame}
\usage{
data( copynumber.data.chr18 )
}
\description{
A test dataset containing chr18 copynumber data
 for the BreakPointGenes package. For more info about the
 CAIRO study see...
}
\keyword{datasets}

