% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{copynumber.data.chr18}
\alias{copynumber.data.chr18}
\title{CGHcall Example copynumber data hg18 chr18}
\format{An object of class cghCall}
\usage{
data( copynumber.data.chr18 )
}
\value{
An object of class cghCall
}
\description{
A test dataset containing copynumber data of chromosome 18 for the GeneBreak package (hg18 based).
}
\keyword{datasets}

