% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{copynumber.data.chr20}
\alias{copynumber.data.chr20}
\title{CGHcall CAIRO data hg19 chr20}
\format{A data.frame}
\usage{
data( copynumber.data.chr20 )
}
\description{
A test dataset containing chr20 copynumber data
 for the BreakPointGenes package. For more info about the
 CAIRO study see...
}
\keyword{datasets}

