% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{copynumber.data.chr21}
\alias{copynumber.data.chr21}
\title{CGHcall CAIRO data hg18 chr21}
\format{A data.frame}
\usage{
data( copynumber.data.chr21 )
}
\description{
A test dataset containing chr21 copynumber data
 for the BreakPointGenes package.
}
\details{
For more info about the CAIRO study see...
}
\keyword{datasets}

