% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{copynumber.data.chr21}
\alias{copynumber.data.chr21}
\title{CGHcall Example copynumber data hg18 chr21}
\format{An object of class cghCall}
\usage{
data( copynumber.data.chr21 )
}
\value{
An object of class cghCall
}
\description{
A test dataset containing chromosome 21 copynumber data for the GeneBreak package (hg18 based). This copy number aberration (CNA) data was obtained by analysis of 200 array-CGH (Agilent 180k) samples from advanced colorectal cancers.
}
\keyword{datasets}

