% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CopyNumberBreakPointGenes-class}
\alias{CopyNumberBreakPointGenes}
\alias{CopyNumberBreakPointGenes-class}
\title{An S4 class to represent a CopyNumberBreakPointGenes object}
\description{
An S4 class to represent a CopyNumberBreakPointGenes object
}
\section{Slots}{

\describe{
\item{\code{geneAnnotation}}{A data.frame with originally given gene information}

\item{\code{geneData}}{A data.frame with gene information added by package methods}

\item{\code{featuresPerGene}}{A list with per gene the associated features}

\item{\code{breakpointsPerGene}}{A matrix with breakage status per gene}
}}

