% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CopyNumberBreakPoints-class}
\alias{CopyNumberBreakPoints}
\alias{CopyNumberBreakPoints-class}
\title{An S4 class to represent a CopyNumberBreakPoints object.}
\description{
An S4 class to represent a CopyNumberBreakPoints object.
}
\section{Slots}{

\describe{
\item{\code{segmDiff}}{A matrix}

\item{\code{callDiff}}{A matrix}

\item{\code{segments}}{A matrix}

\item{\code{calls}}{A matrix}

\item{\code{featureAnnotation}}{A dataframe}

\item{\code{featureData}}{A dataframe}
}}

