% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{CopyNumberBreakPoints-class}
\alias{CopyNumberBreakPoints}
\alias{CopyNumberBreakPoints-class}
\title{An S4 class to represent a CopyNumberBreakPointGenes object.}
\description{
An S4 class to represent a CopyNumberBreakPointGenes object.
}
\section{Slots}{

\describe{
\item{\code{segDiff}}{A matrix}

\item{\code{callDiff}}{A matrix}

\item{\code{segments}}{A matrix}

\item{\code{calls}}{A matrix}

\item{\code{annotation}}{A dataframe}
}}

