% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{featureInfo,CopyNumberBreakPoints-method}
\alias{featureInfo}
\alias{featureInfo,CopyNumberBreakPoints-method}
\title{Access Options featureInfo. This method returns a dataframe with feature annotations.}
\usage{
\S4method{featureInfo}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{of class \code{CopyNumberBreakPoints}}
}
\value{
data.frame with feature annotations
}
\description{
Access Options featureInfo. This method returns a dataframe with feature annotations.
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )
bp <- bpGenes( bp )

featureInfo( bp )
}

