% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{geneInfo,CopyNumberBreakPointGenes-method}
\alias{geneInfo}
\alias{geneInfo,CopyNumberBreakPointGenes-method}
\title{Access Options geneInfo. This method returns a dataframe with gene annotations.}
\usage{
\S4method{geneInfo}{CopyNumberBreakPointGenes}(object)
}
\arguments{
\item{object}{of class \code{CopyNumberBreakPointGenes}}
}
\value{
data.frame with gene annotations
}
\description{
Access Options geneInfo. This method returns a dataframe with gene annotations.
}
\examples{
data( copynumber.data.chr20 )
data( ens.gene.ann.hg18 )
bp <- getBreakpoints( copynumber.data.chr20 )
bp <- bpFilter( bp )
bp <- addGeneAnnotation( bp, ens.gene.ann.hg18 )

geneInfo( bp )
}

