% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BreakPointGenesMethods.R
\name{getBreakpoints}
\alias{getBreakpoints}
\title{CopyNumber to BreakPoints}
\usage{
getBreakpoints(data, first.rm = TRUE)
}
\arguments{
\item{data}{A "CGHcall" object}

\item{first.rm}{Remove the first}
}
\value{
Object of class \code{CopyNumberBreakPoints}.
}
\description{
Builds the CopyNumberBreakPoints object with CGHcall data
}
\examples{
getBreakpoints( cghCallObj )
}

