% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneBreakAccessors.R
\docType{methods}
\name{sampleNames,CopyNumberBreakPoints-method}
\alias{sampleNames}
\alias{sampleNames,CopyNumberBreakPoints-method}
\title{Access Object sampleNames. This method returns a vector with sample names.}
\usage{
\S4method{sampleNames}{CopyNumberBreakPoints}(object)
}
\arguments{
\item{object}{An object of class \code{CopyNumberBreakPoints}}
}
\value{
a vector with sample names
}
\description{
Access Object sampleNames. This method returns a vector with sample names.
}
\examples{
data( copynumber.data.chr20 )
bp <- getBreakpoints( copynumber.data.chr20 )

sampleNames( bp )
}

